Move the core builder-transform logic into its own file.
